backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary       # use for PostgreSQL, for MySQL: mysqlclient
beautifulsoup4
requests
langchain-core
langchain-community
langchain-google-genai
python-dotenv
